Changed test filenames so tests can run, will change them back once I get time to modify the test runner
